MCR-1717 fix command name and moved fix MCR-1717 to migration commands
